FilesystemLoader: added unit test to make that baseDirs with a protocol function properly.
The test itself creates a temporary file, converts it to a directory and then copies two templates over to the directly. It then accesses those files with the file:// php protocol.

Cleanup of temp files (and directory) is done in tearDownAfterClass instead of in the function so that resources are still cleaned up in the event that the test fails/breaks during runtime.