Introduce a strict_variables option to the Engine
When the strict_variables option is enabled, any encountered
unknown variables throw an UnknownVariableException.

This option is useful in a development/CI environnement to catch
mistakes in templates early instead of silently returning an empty string.

This kind of behavior in case of a variable "miss" is accepted by the
mustache manual [1].

[1] https://mustache.github.io/mustache.5.html#Variables