Fix one more php74 notice
Surely we also can use !empty() is isset() bus as far
as this is explicitly about arrays I decided to use that check.

Note this sums on #349 that also is 100% needed.

Trying to use values of type null, bool, int, float or resource as an
array (such as $null["key"]) will now generate a notice. This does not
affect array accesses performed by list().
RFC: https://wiki.php.net/rfc/notice-for-non-valid-array-container